+++ database/factories/OrderFactory.php
@@ -9,8 +9,7 @@ $factory->define(Order::class, function (Faker $faker) {
     return [
         'code' => $faker->regexify('[A-Z0-9]{6}'),
         'email' => $faker->unique()->safeEmail,
-        'shipping_method_id' => rand(1, 4),
-        'shipping_price' => rand(1, 9) + 10.99,
+        'currency' => $faker->currencyCode,
         'status_id' => rand(1, 3),
     ];
 });